requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.3
PyJWT==2.10.0
PyMySQL==1.1.1
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.44.0
whitenoise==6.8.2
